move types around
