+++ apps/nestjs-wallet/src/vc-api/credentials/types/credential.ts
