+++ apps/nestjs-wallet/src/vc-api/credentials/types/verifiable-credential.ts
