+++ apps/nestjs-wallet/src/vc-api/credentials/types/verify-options.ts
@@ -1,0 +1,9 @@
+/**
+ * See "VerifyOptions" from
+ * - https://w3c-ccg.github.io/vc-api/verifier.html#operation/verifyCredential
+ * - https://w3c-ccg.github.io/vc-api/verifier.html#operation/verifyPresentation
+ */
+export interface VerifyOptions {
+  challenge?: string;
+  proofPurpose?: string;
+}
